+++ neuroscout/populate/__init__.py
@@ -1,10 +1,9 @@
 """ Database population methods """
 
 from .extract import extract_features
-from .ingest import (ingest_from_json, add_predictor,
-                     add_group_predictors, add_task)
+from .ingest import add_predictor, add_group_predictors, add_task
 from .modify import delete_task
-
+from .json import ingest_from_json
 
 __all__ = [
     'add_group_predictors',
