+++ neuroscout/populate/ingest.py
@@ -5,7 +5,6 @@ from flask import current_app
 from os.path import realpath, join
 from os import makedirs
 import json
-from pathlib import Path
 import magic
 
 import pandas as pd
@@ -20,8 +19,6 @@ from .utils import remote_resource_exists, format_preproc, hash_file
 from models import (Dataset, Task, Run, Predictor, PredictorEvent, PredictorRun,
                     Stimulus, RunStimulus, GroupPredictor, GroupPredictorValue)
 
-import populate
-
 # from pliers.updater import check_updates
 
 def check_updates_json():
@@ -32,57 +29,6 @@ def check_updates_json():
     updated_extractors = []
 
     return updated_extractors
-
-def ingest_from_json(db_session, config_file, install_path='/file-data',
-                     automagic=False, update=False):
-    dataset_config = json.load(open(config_file, 'r'))
-
-    """ Loop over each dataset in config file """
-    dataset_ids = []
-    for dataset_name, items in dataset_config.items():
-        dataset_address = items.get('dataset_address')
-        preproc_address = items.get('preproc_address')
-        local_path = items.get('path')
-        if not (local_path or dataset_address):
-            raise Exception("Must provide path or remote address to dataset.")
-
-        # If dataset is remote link, set install path
-        if local_path is None:
-            new_path = str((Path(install_path) / dataset_name).absolute())
-        else:
-            new_path = local_path
-
-        for task_name, params in items['tasks'].items():
-            """ Add task to database"""
-            dp = dict(params.get('filters', {}).items() | \
-                      params.get('dataset_args', {}).items())
-
-            dataset_id = add_task(db_session, task_name,
-                                  dataset_name=dataset_name,
-                                  local_path=local_path,
-                                  dataset_address=dataset_address,
-                                  automagic=automagic,
-                                  install_path=new_path,
-                                  preproc_address=preproc_address,
-            					  **dp)
-            dataset_ids.append(dataset_id)
-            dataset_name = Dataset.query.filter_by(id=dataset_id).one().name
-
-            """ Convert stimuli """
-            automagic = local_path is None or automagic
-            converters = params.get('converters', {})
-            if converters:
-                populate.convert_stimuli(db_session, dataset_name, task_name,
-                                         converters, automagic=automagic)
-
-            """ Extract features from applicable stimuli """
-            extractors = params.get('extractors', {})
-            if extractors:
-                populate.extract_features(db_session, dataset_name, task_name,
-                                          extractors, automagic=automagic,
-                                          **params.get('extract_args',{}))
-
-    return dataset_ids
 
 def add_predictor(db_session, predictor_name, dataset_id, run_id,
                   onsets, durations, values, **kwargs):
@@ -310,6 +256,9 @@ def add_task(db_session, task_name, dataset_name=None, local_path=None,
                 path = join(local_path, 'stimuli/{}'.format(val))
 
                 try:
+                    if automagic:
+                        dl.get(path)
+                        dl.unlock(path)
                     stim_hash = hash_file(path)
                 except OSError:
                     print('Stimulus: {} not found. Skipping.'.format(val))
